requirements.txt
langchain==0.0.299
matplotlib==3.7.1
numpy==1.24.3
openai==0.28.0
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.7.0.72
# paddleocr==2.6.1.3
pdf2image==1.16.3
pdfminer.six==20221105
PyPDF2==3.0.1
Requests==2.31.0
streamlit==1.26.0
# paddlepaddle==2.4.2
